Copied from Perforce
 Change: 192192
 ServerID: perforce.ravenbrook.com

+++ mps/test/test/testlib/lofmt.h
@@ -56,7 +56,7 @@ locell *alloclo(mps_ap_t ap, size_t bytes);
 
 long int getloid(locell *obj);
 long int getlocopycount(locell *obj);
-long int getlosize(locell *obj);
+size_t getlosize(locell *obj);
 
 #endif
